+++ Unity/VR Adventure/Assets/Scripts/Networking/Client/EmulatedFireball.cs
@@ -18,7 +18,7 @@ public class EmulatedFireball : MonoBehaviour {
     private Vector3 _newPos;
     private Vector3 _velocity;
 
-    public int _lastUpdateTick = 0;
+    public long _lastUpdateTick = 0;
 
     private DateTime _lastTick;
     private float _tickTime;
